[e2e] move CreateTestOutputDir to common
Remove dependency from client to e2e
